Rename some godot-ffi symbols
